requirements.txt
Pillow>=11.2.1
numpy>=2.2.6
huggingface_hub>=0.32.2
safetensors>=0.5.3
matplotlib>=3.10.3
gradio>=5.31.0
